refactor: renomeia arquivos e componentes de filtros
