+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/PaginatedTable/modules/DataTable/index.tsx
@@ -1,12 +1,10 @@
 'use client'
-import type { GridPaginationModel } from "@mui/x-data-grid";
-import type { Dispatch, SetStateAction, } from "react";
 import { tableDataHook } from "./DataTable.hook";
 import { diabetesColumns } from "./modules/diabetes/modules/columns/columns";
 import { Table } from "@impulsogov/design-system";
 import { EmptyTableMessage } from "./modules/EmptyTableMessage";
 import { useContext } from 'react';
-import { FiltersContext } from "../../../TableWithFilters/TableWithFilters.context";
+import { FiltersContext } from "../../../WithFilters/context";
 import type { SelectedValues } from "@/features/acf/diabetes/frontend/model";
 import { PaginationContext } from "../../PaginatedTable.context";
 import type { PaginationModel } from "../..";
