+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/print/buildPrintProps.ts
@@ -2,7 +2,7 @@ import type { PrintTableProps } from "@/componentes/unmounted/lista-nominal/prin
 import { larguraColunasHipertensaoEquipePaisagem, larguraColunasHipertensaoEquipeRetrato, larguraColunasHipertensaoPaisagem, larguraColunasHipertensaoRetrato } from "@/helpers/larguraColunasHipertensao";
 import type { FilterItem } from "@/services/lista-nominal/ListaNominal";
 import { PROFILE_ID, type ProfileIdValue } from "@/types/profile";
-import { filtersLabels } from "../TableWithFilters/modules/filters/modules/diabetes/filtersLabels";
+import { filtersLabels } from "../WithFilters/modules/filtersLabels";
 import type { DiabetesAcfPrintItem } from "./diabetes/print.model";
 import { diabetesColumns } from "../PaginatedTable/modules/DataTable/modules/diabetes/modules/columns/columns";
 import type { AcfDashboardType } from "@/features/acf/diabetes/common/model";
